deadcode linter fixes
Changes:
- Enables deadcode linter to ci scripts
- Approaches used to fix lint errors
    - Deletes unused code
    - Co-locates definition and calls wrt build tags.
    - Adds //nolint:deadcode exemptions with explanations.
- Addresses Issue-99243
