+++ cmd/kubelet/app/options/globalflags.go
@@ -51,17 +51,6 @@ func normalize(s string) string {
 	return strings.Replace(s, "_", "-", -1)
 }
 
-// register adds a flag to local that targets the Value associated with the Flag named globalName in global
-func register(global *flag.FlagSet, local *pflag.FlagSet, globalName string) {
-	if f := global.Lookup(globalName); f != nil {
-		pflagFlag := pflag.PFlagFromGoFlag(f)
-		pflagFlag.Name = normalize(pflagFlag.Name)
-		local.AddFlag(pflagFlag)
-	} else {
-		panic(fmt.Sprintf("failed to find flag in global flagset (flag): %s", globalName))
-	}
-}
-
 // pflagRegister adds a flag to local that targets the Value associated with the Flag named globalName in global
 func pflagRegister(global, local *pflag.FlagSet, globalName string) {
 	if f := global.Lookup(globalName); f != nil {
@@ -70,12 +59,6 @@ func pflagRegister(global, local *pflag.FlagSet, globalName string) {
 	} else {
 		panic(fmt.Sprintf("failed to find flag in global flagset (pflag): %s", globalName))
 	}
-}
-
-// registerDeprecated registers the flag with register, and then marks it deprecated
-func registerDeprecated(global *flag.FlagSet, local *pflag.FlagSet, globalName, deprecated string) {
-	register(global, local, globalName)
-	local.Lookup(normalize(globalName)).Deprecated = deprecated
 }
 
 // addCredentialProviderFlags adds flags from k8s.io/kubernetes/pkg/credentialprovider
