+++ pkg/kubelet/config/file.go
@@ -37,10 +37,6 @@ import (
 type podEventType int
 
 const (
-	podAdd podEventType = iota
-	podModify
-	podDelete
-
 	eventBufferLen = 10
 )
