+++ pkg/kubelet/dockershim/docker_service.go
@@ -61,8 +61,6 @@ const (
 	// '{{.HostConfig.NetworkMode}}'.
 	namespaceModeHost = "host"
 
-	dockerNetNSFmt = "/proc/%v/ns/net"
-
 	// Internal docker labels used to identify whether a container is a sandbox
 	// or a regular container.
 	// TODO: This is not backward compatible with older containers. We will
