+++ pkg/registry/core/pod/rest/metrics.go
@@ -29,7 +29,6 @@ const (
 
 	usageEnforce     = "enforce_tls"
 	usageSkipAllowed = "skip_tls_allowed"
-	usageSkipDenied  = "skip_tls_denied"
 )
 
 var (
