+++ pkg/util/iptables/iptables_unsupported.go
@@ -20,13 +20,8 @@ package iptables
 
 import (
 	"fmt"
-	"os"
 )
 
 func grabIptablesLocks(lock14filePath, lock16filePath string) (iptablesLocker, error) {
 	return nil, fmt.Errorf("iptables unsupported on this platform")
-}
-
-func grabIptablesFileLock(f *os.File) error {
-	return fmt.Errorf("iptables unsupported on this platform")
 }
