+++ pkg/volume/awsebs/aws_util.go
@@ -20,8 +20,6 @@ package awsebs
 
 import (
 	"fmt"
-	"os"
-	"path/filepath"
 	"strconv"
 	"strings"
 	"time"
@@ -41,9 +39,7 @@ import (
 )
 
 const (
-	diskPartitionSuffix     = ""
-	nvmeDiskPartitionSuffix = "p"
-	checkSleepDuration      = time.Second
+	checkSleepDuration = time.Second
 )
 
 // AWSDiskUtil provides operations for EBS volume.
@@ -201,6 +197,7 @@ func populateVolumeOptions(pluginName, pvcName string, capacityGB resource.Quant
 }
 
 // Returns the first path that exists, or empty string if none exist.
+// nolint:deadcode
 func verifyDevicePath(devicePaths []string) (string, error) {
 	for _, path := range devicePaths {
 		if pathExists, err := mount.PathExists(path); err != nil {
@@ -213,44 +210,6 @@ func verifyDevicePath(devicePaths []string) (string, error) {
 	return "", nil
 }
 
-// Returns list of all paths for given EBS mount
-// This is more interesting on GCE (where we are able to identify volumes under /dev/disk-by-id)
-// Here it is mostly about applying the partition path
-func getDiskByIDPaths(volumeID aws.KubernetesVolumeID, partition string, devicePath string) []string {
-	devicePaths := []string{}
-	if devicePath != "" {
-		devicePaths = append(devicePaths, devicePath)
-	}
-
-	if partition != "" {
-		for i, path := range devicePaths {
-			devicePaths[i] = path + diskPartitionSuffix + partition
-		}
-	}
-
-	// We need to find NVME volumes, which are mounted on a "random" nvme path ("/dev/nvme0n1"),
-	// and we have to get the volume id from the nvme interface
-	awsVolumeID, err := volumeID.MapToAWSVolumeID()
-	if err != nil {
-		klog.Warningf("error mapping volume %q to AWS volume: %v", volumeID, err)
-	} else {
-		// This is the magic name on which AWS presents NVME devices under /dev/disk/by-id/
-		// For example, vol-0fab1d5e3f72a5e23 creates a symlink at /dev/disk/by-id/nvme-Amazon_Elastic_Block_Store_vol0fab1d5e3f72a5e23
-		nvmeName := "nvme-Amazon_Elastic_Block_Store_" + strings.Replace(string(awsVolumeID), "-", "", -1)
-		nvmePath, err := findNvmeVolume(nvmeName)
-		if err != nil {
-			klog.Warningf("error looking for nvme volume %q: %v", volumeID, err)
-		} else if nvmePath != "" {
-			if partition != "" {
-				nvmePath = nvmePath + nvmeDiskPartitionSuffix + partition
-			}
-			devicePaths = append(devicePaths, nvmePath)
-		}
-	}
-
-	return devicePaths
-}
-
 // Return cloud provider
 func getCloudProvider(cloudProvider cloudprovider.Interface) (*aws.Cloud, error) {
 	awsCloudProvider, ok := cloudProvider.(*aws.Cloud)
@@ -259,38 +218,6 @@ func getCloudProvider(cloudProvider cloudprovider.Interface) (*aws.Cloud, error)
 	}
 
 	return awsCloudProvider, nil
-}
-
-// findNvmeVolume looks for the nvme volume with the specified name
-// It follows the symlink (if it exists) and returns the absolute path to the device
-func findNvmeVolume(findName string) (device string, err error) {
-	p := filepath.Join("/dev/disk/by-id/", findName)
-	stat, err := os.Lstat(p)
-	if err != nil {
-		if os.IsNotExist(err) {
-			klog.V(6).Infof("nvme path not found %q", p)
-			return "", nil
-		}
-		return "", fmt.Errorf("error getting stat of %q: %v", p, err)
-	}
-
-	if stat.Mode()&os.ModeSymlink != os.ModeSymlink {
-		klog.Warningf("nvme file %q found, but was not a symlink", p)
-		return "", nil
-	}
-
-	// Find the target, resolving to an absolute path
-	// For example, /dev/disk/by-id/nvme-Amazon_Elastic_Block_Store_vol0fab1d5e3f72a5e23 -> ../../nvme2n1
-	resolved, err := filepath.EvalSymlinks(p)
-	if err != nil {
-		return "", fmt.Errorf("error reading target of symlink %q: %v", p, err)
-	}
-
-	if !strings.HasPrefix(resolved, "/dev") {
-		return "", fmt.Errorf("resolved symlink for %q was unexpected: %q", p, resolved)
-	}
-
-	return resolved, nil
 }
 
 func formatVolumeID(volumeID string) (string, error) {
