+++ test/e2e/upgrades/apps/replicasets.go
@@ -19,6 +19,7 @@ package apps
 import (
 	"context"
 	"fmt"
+
 	appsv1 "k8s.io/api/apps/v1"
 	v1 "k8s.io/api/core/v1"
 	metav1 "k8s.io/apimachinery/pkg/apis/meta/v1"
